requirements.txt
Django==1.7
PyYAML==3.11
South==1.0
Werkzeug==0.9.6
argparse==1.2.1
certifi==14.05.14
distribute==0.6.24
django-allauth==0.18.0
django-tastypie==0.12.0
django-werkzeug-debugger-runserver==0.2
djangorestframework==2.4.2
ipython==2.2.0
oauthlib==0.6.3
psycopg2==2.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
requests==2.4.0
requests-oauthlib==0.4.1
six==1.7.3
wsgiref==0.1.
